Cache path values to disk to minimize eval calls
`chruby_use()`'s `PATH`, `GEM_HOME`, and `GEM_PATH` environment variable
setting functionality has been relocated to a new `chruby_set_paths()`
function. This new function will cache to disk the results of the
`eval` call which invokes Ruby to determine the values to use. The
cached results will be looked for and sourced if found in lieu of
invoking the `eval()` call. This logic is one potential way to address
[issue 369](https://github.com/postmodern/chruby/issues/369).